Mipmap support, Compute shaders, some prep for vulkan and maybe move to hlsl

+++ Stulu/src/Stulu/Scene/Components/Camera.cpp
@@ -16,7 +16,7 @@ namespace Stulu {
 			
 		if (genMap) {
 			if (reflectionMap == nullptr)
-				reflectionMap = CubeMap::create(ST_REFLECTION_MAP_SIZE, TextureSettings{ (int)TextureSettings::Format::RGB16F });
+				reflectionMap = CubeMap::create(ST_REFLECTION_MAP_SIZE, TextureSettings{ TextureSettings::Format::RGB16F });
 			if (reflectionFrameBuffer == nullptr)
 				reflectionFrameBuffer = FrameBuffer::create(FrameBufferSpecs{ ST_REFLECTION_MAP_SIZE,ST_REFLECTION_MAP_SIZE,1,false, TextureSettings::Format::RGB16F });
 		}
